requirements.txt
fastapi
uvicorn
pandas
numpy
sentence-transformers
torch
pymongo
python-dotenv
google-generativeai
gtts
